+++ gislib/stores.py
@@ -5,19 +5,11 @@ from __future__ import unicode_literals
 from __future__ import absolute_import
 from __future__ import division
 
-import collections
-import datetime
-import glob
-import logging
 import math
-import os
-import time
-import shutil
 
 from osgeo import gdal
 from osgeo import gdal_array
 from osgeo import ogr
-from osgeo import osr
 
 import numpy as np
 
@@ -26,46 +18,13 @@ from gislib import rasters
 from gislib import utils
 from gislib import vectors
 
+OGR_MEM_DRIVER = ogr.GetDriverByName(b'Memory')
+
 
 class BaseStore(object):
-    """ Base class for anything that has a warpinto method. """
-    
-    def get_array(self, extent, size, projection):
-        """
-        Return data for the pyramids.
-
-        Datatype and nodatavalue are taken from first pyramid.
-
-        extent: xmin, xmax, ymin, ymax-tuple
-        size: width, height-tuple
-        projection: something like 'epsg:28992' or a wkt or proj4 string.
-        """
-        info = self.info
-
-        # Create a dataset
-        array = np.ones(
-            (
-                1,
-                size[1],
-                size[0],
-            ),
-            dtype=gdal_array.flip_code(info['datatype']),
-        ) * info['nodatavalue']
-        dataset = rasters.array2dataset(array)
-
-        # Add georeferencing
-        xmin, ymin, xmax, ymax = extent
-        geotransform = (xmin, (xmax - xmin) / array.shape[-1], 0,
-                        ymax, 0, (ymin - ymax) / array.shape[-2])
-        dataset.SetProjection(projections.get_wkt(projection))
-        dataset.SetGeoTransform(geotransform)
-
-        # Get data
-        self.warpinto(dataset)
-        dataset.FlushCache()
-
-        return np.ma.masked_equal(array, info['nodatavalue'], copy=False)
-
+    """
+    Base class for anything that has a warpinto method.
+    """
     def get_profile(self, line, size, projection):
         """
         Return a distances, values tuple of numpy arrays.
@@ -102,7 +61,7 @@ class BaseStore(object):
         ).transpose())[::-1]
 
         # Get the values from the array
-        array = self.get_array(extent, (width, height), projection)
+        array = self.get_array(extent, (width, height), crs=projection)
         values = array[0][indices]
 
         # make array with distance from origin (x values for graph)
@@ -111,9 +70,91 @@ class BaseStore(object):
 
         return distances, values
 
+    def get_data(self, wkt, crs, size=None):
+        """
+        Generalized data extraction from store interfaces.
+        """
+        wkb = ogr.CreateGeometryFromWkt(wkt)
+        handler = self.HANDLERS[wkb.GetGeometryType()]
+        return handler(self, wkb, crs, size=size)
+
+    def get_data_for_polygon(self, wkb, crs, size):
+        """
+        Return a numpy array for the data.
+        """
+        # Quick out if polygon bounds match polygon
+        geometry = vectors.Geometry(wkb)
+        envelope = geometry.envelope
+        extent = geometry.extent
+        nodatavalue = self.info['nodatavalue']
+        datatype = self.info['datatype']
+
+        # Initialize resulting array to nodatavalue
+        array = np.ones(
+            (1, size[1], size[0]),
+            dtype=gdal_array.flip_code(datatype),
+        ) * nodatavalue
+
+        # Create dataset and use it to retrieve data from the store
+        dataset = rasters.array2dataset(array=array, extent=extent, crs=crs)
+        self.warpinto(dataset)
+        dataset.FlushCache()
+
+        # Cut when necessary
+        if not envelope.Equals(wkb):
+            source = OGR_MEM_DRIVER.CreateDataSource('')
+            sr = projections.get_spatial_reference(crs)
+            layer = source.CreateLayer(b'', sr)
+            defn = layer.GetLayerDefn()
+            feature = ogr.Feature(defn)
+            feature.SetGeometry(wkb)
+            layer.CreateFeature(feature)
+            gdal.RasterizeLayer(dataset, (1,), layer,
+                                burn_values=(nodatavalue,))
+            dataset.FlushCache()
+
+        return np.ma.masked_equal(array, nodatavalue, copy=False)
+
+    def get_data_for_linestring(self, wkb, crs, size):
+        """ What the name says. """ 
+        segmentsize = wkb.Length() / size
+        geometry = vectors.Geometry(wkb)
+        span = geometry.size
+
+        # Cellsize must be such that it fits integer amount in extent
+        cellsize = np.array(tuple(s / (s // segmentsize) for s in span))
+        
+        # Segmentize and extract the points
+        wkb.Segmentize(segmentsize)
+        points = np.fromstring(wkb.ExportToWkb()[9:]).byteswap().reshape(-1, 2)
+        
+        # Get 2D data
+        envelope = geometry.envelope
+        datasize = tuple(int(s / c) for s, c in zip(span, cellsize))
+        data = self.get_data_for_polygon(crs=crs, wkb=envelope, size=datasize)
+
+        # Extract 1D data
+        x1, y1, x2, y2 = geometry.extent
+        origin = np.array([x1, y2])
+        indices = tuple(np.minimum(
+            np.uint64((points - origin) / cellsize * np.array([1, -1])),
+            np.uint64(datasize) - 1,
+        ).transpose())[::-1]
+
+        return data[0][indices]
+
+    def get_data_for_point(self, wkb, crs, size):
+        pass
+
+    HANDLERS = {
+        ogr.wkbPolygon: get_data_for_polygon,
+        ogr.wkbLineString: get_data_for_linestring,
+        ogr.wkbPoint: get_data_for_point,
+    }
+
 
 class MultiStore(BaseStore):
-    """ Pyramid wrapper for data extraction from a list of pyramids. """
+    """ Store wrapper for data extraction from a list of stores. """
     def __init__(self, stores):
         self.stores = stores
 
@@ -121,8 +162,8 @@ class MultiStore(BaseStore):
     def info(self):
         """ Return store info. """
         return self.stores[0].info
-    
+
     def warpinto(self, dataset):
         """ Multistore version of warpinto. """
-        for store in stores:
+        for store in self.stores:
             store.warpinto(dataset)